Feature: Configurable input (#15)
Adds a new section to Options that allows configuration of controls
(looks like shit rn):

![image](https://github.com/user-attachments/assets/f580315a-2017-493f-967c-01b8042f192e)

Allows to configure Keyboard/Mouse bindings in one column and Gamepad
ones in the other. Configuration is stored in `BindingsSettings` and
`Input` is reconfigured each time as necessary.

**Bugs:**
* Tgui scrollable panel doesn't correctly compute height of the options
content so the scrollbar doesn't work at all and the configuration can
bleed out of screen on small screens.
* When "back" button is configured, the Options screen is immediately
exited

**Tasks**
- [x] Deploy on Android, check for rendering errors (gamepads not yet
supported in SFML)
- [x] Try to fix the back issue
- [x] Hide or fix scrollable panel for now